third_party/blink/web_tests/fast/dom/DOMError-constructor-expected.txt

Tests that DOMError is constructible.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS error = new DOMError('NotFoundError', 'This item was not found.') did not throw exception.
PASS error.name is "NotFoundError"
PASS error.message is "This item was not found."
PASS error = new DOMError('IndexSizeError') did not throw exception.
PASS error.name is "IndexSizeError"
PASS error.message is ""
PASS new DOMError() threw exception TypeError: Failed to construct 'DOMError': 1 argument required, but only 0 present..
PASS successfullyParsed is true

TEST COMPLETE
